Move `_StateObject` to "internal/"
